backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain
langchain-core
langgraph
langchain-google-genai
supabase
tavily-python
weasyprint
jinja2
python-dotenv
